refactor: move settings to a separate mod
