+++ src/main.rs
@@ -1,8 +1,7 @@
 use anyhow::Result;
 use clap::{command, CommandFactory, Parser};
 use clap_complete::{generate, Shell};
-use serde_derive::{Deserialize, Serialize};
-use std::{fs, io};
+use std::io;
 
 #[derive(Parser)]
 #[command(name = "VFD-Tools")]
@@ -19,19 +18,8 @@ pub struct CliArguments {
     pub fest: bool,
 }
 
-#[derive(Serialize, Deserialize, Debug)]
-pub struct Settings {
-    pub profiles: Vec<Profile>,
-    #[serde(rename = "vfd-ssh-git")]
-    pub vfd_ssh_git: String,
-}
-#[derive(Serialize, Deserialize, Debug)]
-pub struct Profile {
-    name: String,
-    services: Vec<String>,
-}
-
 mod runner;
+mod settings;
 
 fn main() -> Result<()> {
     let args = CliArguments::parse();
@@ -43,9 +31,6 @@ fn main() -> Result<()> {
         return Ok(());
     }
 
-    let setting_file_contents =
-        fs::read_to_string("./settings.json").expect("Failed to read settings.json");
-    let settings: Settings = serde_json::from_str(setting_file_contents.as_str())?;
-
+    let settings = settings::get_settings();
     runner::Runner::new(args, settings).run()
 }
